Create Code
A Dodger game is a type of game in which a player has access to control an object or any character to dodge (to avoid) upcoming obstacles and earn points according to the number of items dodged if the player is not able to dodge the object the game will be over and final score will be displayed in the console. After certain points or times the speed of the objects to be dodged increases which makes it harder and harder to score more points. The player can control the movement through some input from some input devices, it can a keyboard, a joystick, a game controller, etc..